Move running_builds route to BuildController
The new path is `/builds/running_builds`.
