--- apps/rabbitci_core/web/views/index_view.ex
@@ -1,8 +1,0 @@
-defmodule RabbitCICore.IndexView do
-  use RabbitCICore.Web, :view
-  alias RabbitCICore.BuildSerializer
-
-  def render("running_builds.json", %{conn: conn, builds: builds}) do
-    BuildSerializer.format(builds, conn, %{})
-  end
-end
